Добавил DataManagerImpl, позволяющий общаться с репозиториями, через него идут примитивные CRUD операции с объектами небольшие модификации спецификаций и фабрик

+++ Project/src/main/java/specifications/sql/impl/SeatByLineIdSqlSpecification.java
@@ -11,7 +11,7 @@ public class SeatByLineIdSqlSpecification implements SqlSpecification<Seat> {
     private long lineId;
     private String tableName;
 
-    public SeatByLineIdSqlSpecification(int lineId) {
+    public SeatByLineIdSqlSpecification(long lineId) {
         this.lineId = lineId;
         tableName = DataBaseNames.SEATS;
     }
